realm based library root controller
