+++ iOS/Controller/Library/LibrarySplitController.swift
@@ -1,0 +1,40 @@
+//
+//  LibrarySplitController.swift
+//  iOS
+//
+//  Created by Chris Li on 4/30/18.
+//  Copyright © 2018 Chris Li. All rights reserved.
+//
+
+import UIKit
+
+class LibrarySplitController: UISplitViewController, UISplitViewControllerDelegate {
+    init() {
+        super.init(nibName: nil, bundle: nil)
+        config()
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        config()
+    }
+    
+    private func config() {
+        // set at least one view controller in viewControllers to supress a warning produced by split view controller
+        viewControllers = [UIViewController()]
+        
+        preferredDisplayMode = .allVisible
+        delegate = self
+        
+        let master = LibraryMasterController()
+        let detail = UIViewController()
+        detail.view.backgroundColor = .groupTableViewBackground
+        viewControllers = [
+            UINavigationController(rootViewController: master),
+            UINavigationController(rootViewController: detail)]
+    }
+    
+    func splitViewController(_ splitViewController: UISplitViewController, collapseSecondary secondaryViewController: UIViewController, onto primaryViewController: UIViewController) -> Bool {
+        return true
+    }
+}
